+++ gwwnetworktrace/gww_gis_tools/trace_gis/trace_sewer.py
@@ -1,64 +1,92 @@
+"""Trace Sewer."""
+
 from __future__ import annotations
 
+import builtins
+import contextlib
 import inspect
 import json
 from collections import defaultdict
 from enum import Enum
-from typing import Callable, Union
+from pathlib import Path
+from typing import TYPE_CHECKING, Callable
 
-try:
-    from typing import Self
-except ImportError:
-    from typing_extensions import Self
+with contextlib.suppress(ImportError):
+    from typing import Any, Self
+
+if TYPE_CHECKING:
+    with contextlib.suppress(ImportError):
+        import geopandas as gpd  # pyright: ignore[reportMissingImports]
 
 
 class DIRECTION(Enum):
-    U = "upstream"
-    D = "downstream"
+    """Specify direction on initialisation."""
+
+    U = 'upstream'
+    D = 'downstream'
 
 
 class Graph:
-    """
-    Represents network as a graph data structure.
+    """Represents network as a graph data structure.
 
     Args:
-    - direction: The direction of the graph (either 'U' for undirected or 'D' for directed).
+    ----
+    - direction: The direction of the graph
+        (either 'U' for undirected or 'D' for directed).
 
     Attributes:
+    ----------
     - direction: The direction of the graph.
     - nodes: A dictionary of nodes and their adjacent nodes.
     - pipes: A dictionary of nodes and the pipes associated with them.
 
     Methods:
+    -------
     - _validate_direction(direction): Validates the input direction.
-    - from_gdf(links: geopandas.GeoDataFrame | pandas.DataFrame): Converts a (Geo)DataFrame to a graph by adding edges.
-    - from_dicts(links: list[dict]): Converts a list of dictionaries to a graph by adding edges.
-    - add_edge(start_node, end_node, pipe_id): Adds an edge to the graph based on the direction.
+    - from_gdf(links: geopandas.GeoDataFrame | pandas.DataFrame):
+        Converts a (Geo)DataFrame to a graph by adding edges.
+    - from_dicts(links: list[dict]):
+        Converts a list of dictionaries to a graph by adding edges.
+    - add_edge(start_node, end_node, pipe_id):
+        Adds an edge to the graph based on the direction.
+
     """
 
-    def __init__(self, direction: DIRECTION):
-        """Specify direction on initialisation"""
+    def __init__(self, direction: DIRECTION) -> None:
+        """Specify direction on initialisation."""
         self.direction = direction
-        assert type(self.direction) is DIRECTION
-        self.nodes: defaultdict[list] = defaultdict(list)
-        self.pipes: defaultdict[list] = defaultdict(list)
+
+        self.nodes: defaultdict[str | int, list] = defaultdict(list)
+        self.pipes: defaultdict[str | int, list] = defaultdict(list)
+        self.qgis_fids: dict = {}
 
     def __repr__(self) -> str:
-        return f"Graph({self.direction}, {len(self.nodes)=})"
+        """Return a string representation of the graph."""
+        return f'Graph({self.direction}, {len(self.nodes)=})'
 
-    def from_gdf(self, links) -> Self:
-        """Takes a (Geo)DataFrame and add each row as an edge. Returns graph object."""
+    def from_gdf(self, links: gpd.GeoDataFrame) -> Self:
+        """Use a (Geo)DataFrame to add each row as an edge. Returns graph object."""
         for f in links.itertuples(index=False):
             self.add_edge(f.START_NODE, f.END_NODE, f.PIPE_ID)
         return self
 
     def from_dicts(self, links: list[dict]) -> Self:
-        """Takes a list of dictionaries and add each row as an edge. Returns graph object."""
-        for f in links:
-            self.add_edge(f["START_NODE"], f["END_NODE"], f["PIPE_ID"])
+        """Use a list of dictionaries to add rows as an edges. Returns graph object."""
+        SUPPORTED_KEYS = ['START_NODE', 'END_NODE', 'PIPE_ID', 'QGIS_FID']
+        fields = next(iter(links)).keys()
+        keys = [f for f in SUPPORTED_KEYS if f in fields]
+
+        for link in links:
+            self.add_edge(*tuple(link[k] for k in keys))
         return self
 
-    def add_edge(self, start_node, end_node, pipe_id) -> Self:
+    def add_edge(
+        self,
+        start_node: str | int,
+        end_node: str | int,
+        pipe_id: str | int,
+        qgis_fid: int | None = None
+    ) -> Self:
         """Add field values for a single feature as an edge. Returns graph object."""
         if self.direction == DIRECTION.U:
             self.nodes[end_node].append(start_node)
@@ -66,56 +94,78 @@ class Graph:
         elif self.direction == DIRECTION.D:
             self.nodes[start_node].append(end_node)
             self.pipes[start_node].append(pipe_id)
-        else:
-            msg = f"Graph direction {self.direction} is not valid. Use trace_sewer.DIRECTION.U or DIRECTION.D"
-            raise NotImplementedError(msg)
+
+        if qgis_fid:
+            self.qgis_fids[pipe_id] = qgis_fid
 
         return self
 
-    def to_file(self, filename: str):
-        with open(filename, "w") as f:
+    def to_file(self, filename: str) -> None:
+        """Write graph object to file."""
+        with Path(filename).open('w') as f:
             json.dump(self, f, cls=ExtendedEncoder, sort_keys=True)
 
     @classmethod
     def from_file(cls, filename: str) -> Self:
-        with open(filename) as f:
+        """Read graph object from file."""
+        with Path(filename).open() as f:
             return json.load(f, cls=ExtendedDecoder)
 
 
 class TraceResult:
-    """
-    Dataclass-like objects for accessing results of tracing.
-    """
+    """Dataclass-like objects for accessing results of tracing."""
 
-    def __init__(self, trace_summary, pipes: set, nodes: set, end_of_path_nodes: set):
+    def __init__(
+        self,
+        trace_summary: dict,
+        pipes: set,
+        nodes: set,
+        end_of_path_nodes: set,
+    ) -> None:
+        """Initialize TraceResult."""
         self.trace_summary = trace_summary
         self.pipes = pipes
         self.nodes = nodes
         self.end_of_path_nodes = end_of_path_nodes
 
-    def __repr__(self):
-        return f"TraceResult({self.trace_summary['direction']}, {self.trace_summary['start_node']})"
+    def __repr__(self) -> str:
+        """Return string representation of TraceResult."""
+        return 'TraceResult({}, {})'.format(
+            self.trace_summary['direction'],
+            self.trace_summary['start_node'],
+        )
 
 
 class Trace:
-    """
-    This class is used to trace a path through a graph. Option to provide a function that returns False to stop tracing.
+    """Used to trace a path through a graph.
+
+    Option to provide a function that returns False to stop tracing.
 
     Args:
+    ----
     - graph: The graph to be traced.
     - stop_node: Optional function to determine when to stop tracing.
 
     Methods:
+    -------
     - trace(first_node): Traces a path through the graph starting from the first node.
     """
 
-    def __init__(self, graph, stop_node: Callable | None = None):
+    def __init__(self, graph: Graph, stop_node: Callable | None = None) -> None:
+        """Initialise Trace."""
         self.graph = graph
-        self.stop_node = stop_node or (lambda x: True)
+        self.stop_node = stop_node or (lambda x: True) # noqa: ARG005
 
-    def trace(self, first_node, trace_name=None, summary=False):
+    def trace(
+        self,
+        first_node: str | int,
+        trace_name: str = '',
+        summary: bool = False  # noqa: FBT001, FBT002
+    ) -> TraceResult:
         """
-        THis method tranverses graph object (depth-first) and returns a TraceResult
+        Main trace method.
+
+        This method tranverses graph object (depth-first) and returns a TraceResult
         containing nodes and pipes visited, and end of path nodes.
         """
         nodes = self.graph.nodes
@@ -133,7 +183,8 @@ class Trace:
                 nodes_visited.add(next_node)
                 end_of_path_nodes.add(next_node)
                 break
-            elif next_node not in nodes_visited:
+
+            if next_node not in nodes_visited:
                 nodes_visited.add(next_node)
                 next_nodes = nodes[next_node]
                 next_pipes = pipes[next_node]
@@ -146,14 +197,14 @@ class Trace:
 
         if summary:
             trace_summary = {
-                "trace_name": trace_name,
-                "g_size": len(self.graph.nodes),
-                "direction": self.graph.direction,
-                "start_node": first_node,
-                "stop_node_predicate": str(inspect.getsource(self.stop_node)).strip(),
+                'trace_name': trace_name,
+                'g_size': len(self.graph.nodes),
+                'direction': self.graph.direction,
+                'start_node': first_node,
+                'stop_node_predicate': str(inspect.getsource(self.stop_node)).strip(),
             }
         else:
-            trace_summary = None
+            trace_summary = {}
 
         return TraceResult(
             trace_summary=trace_summary,
@@ -164,76 +215,120 @@ class Trace:
 
 
 class ExtendedEncoder(json.JSONEncoder):
-    """
-    Extends JSONEncoder to customize JSON serialization.
+    """Extends JSONEncoder to customize JSON serialization.
 
     Methods:
-    - default(self, obj): Overrides default behavior for serializing objects and uses dynamic method lookup to encode the object.
-    - encode_TraceResult(self, tr) -> dict[str, list]: Specifically encodes TraceResult objects into a dictionary.
+    -------
+    - default(self, obj): Overrides default behavior for serializing objects
+        and uses dynamic method lookup to encode the object.
+    - encode_TraceResult(self, tr) -> dict[str, list]: Specifically encodes
+        TraceResult objects into a dictionary.
+
     """
 
-    def default(self, obj):
-        name = obj.__class__.__name__
+    def default(self, o: Any) -> str: # noqa: ANN401
+        # TODO @timothy-holmes: create JSON type list # noqa: FIX002, TD003
+        """
+        Overrides default behavior for serializing objects.
+
+        Uses dynamic method lookup to encode the object.
+        """
+        name = o.__class__.__name__
 
         try:
-            encoder = getattr(self, f"encode_{name}")
+            encoder = getattr(self, f'_encode_{name}')
         except AttributeError:
-            super().default(obj)
+            return o
         else:
-            encoded = encoder(obj)
-            encoded["__extended_json_type__"] = name
+            encoded = encoder(o)
+            encoded['__extended_json_type__'] = name
             return encoded
 
-    def encode_defaultdict(self, obj):
+    def _encode_defaultdict(self, o: defaultdict) -> dict[str, str | dict]:
         return {
-            "__extended_json_type__": "defaultdict",
-            "__default_factory__": obj.default_factory.__class__.__name__,
-            "__dict__": dict(obj),
+            '__default_factory__': self.default(o.default_factory),
+            '__dict__': dict(o),
         }
 
-    def encode_TraceResult(self, tr: TraceResult) -> dict[str, list]:  # noqa N802
+    def _encode_type(self, o: type) -> dict[str, str]:
+        return {'_type': o.__name__}
+
+    def _encode_TraceResult(self, tr: TraceResult) -> dict[str, list]:  # noqa: N802
         raise NotImplementedError
 
-    def encode_DIRECTION(self, d: DIRECTION) -> dict[str, str]:  # noqa N802
-        return {"__extended_json_type__": "DIRECTION", "_direction": d.value}
+    def _encode_DIRECTION(self, d: DIRECTION) -> dict[str, str]:  # noqa: N802
+        return {'_direction': d.value}
 
-    def encode_Graph(self, g: Graph) -> dict[str, dict | list | str]:  # noqa N802
+    def _encode_Graph(self, g: Graph) -> dict[str, dict | list | str]:  # noqa: N802
         return {
-            "__extended_json_type__": "Graph",
-            "_direction": self.default(g.direction),
-            "_nodes": g.nodes,
-            "_pipes": g.pipes,
+            '_direction': self.default(g.direction),
+            '_nodes': self.default(g.nodes),
+            '_pipes': self.default(g.pipes),
+            '_qgis_fids': self.default(g.qgis_fids)
         }
 
 
 class ExtendedDecoder(json.JSONDecoder):
-    def __init__(self, **kwargs):
-        kwargs["object_hook"] = self.object_hook
+    """
+    Extends JSONDecoder to customize JSON deserialization.
+
+    Methods:
+    -------
+    - object_hook(self, obj): Overrides default behavior for deserializing objects
+        and uses dynamic method lookup to decode the object.
+    """
+
+    def __init__(self, **kwargs) -> None: # noqa: ANN003
+        """Initialize the class."""
+        kwargs['object_hook'] = self.object_hook
         super().__init__(**kwargs)
 
-    def object_hook(self, obj):
+    def object_hook( # pyright: ignore[reportIncompatibleMethodOverride]
+        self,
+        obj: dict,
+    ) -> Any: # noqa: ANN401
+        """
+        Replace method for extended JSON deserialization.
+
+        Uses dynamic method lookup to decode the object.
+        """
         try:
-            name = obj["__extended_json_type__"]
-            decoder = getattr(self, f"decode_{name}")
-        except (KeyError, AttributeError, TypeError) as e:
+            name = obj['__extended_json_type__']
+            decoder = getattr(self, f'_decode_{name}')
+        except (KeyError, AttributeError, TypeError):
             return obj
         else:
             return decoder(obj)
 
-    def decode_TraceResult(self, tr: TraceResult) -> dict[str, list]:  # noqa N802
+    def _decode_TraceResult(self, obj: dict) -> TraceResult:  # noqa: N802
         raise NotImplementedError
 
-    def decode_tuple(self, obj):
-        return tuple(self.object_hook(element) for element in obj["value"])
+    def _decode_defaultdict(self, obj: dict) -> defaultdict:
+        return defaultdict(
+            self.object_hook(obj['__default_factory__']),
+            self.object_hook(obj['__dict__'])
+        )
 
-    def decode_defaultdict(self, obj):
-        return defaultdict(obj["__default_factory__"], map=self.object_hook(obj["__dict__"]))
+    def _decode_type(self, obj: dict) -> type:  # noqa: N802
+        possible_types = {
+            getattr(builtins,b).__name__: getattr(builtins,b)
+            for b in dir(builtins)
+            if type(getattr(builtins,b)) is type
+        }
+        possible_types |= {type(g).__name__: g for g in globals() if type(g) is type}
+        return possible_types[obj['_type']]
 
-    def decode_DIRECTION(self, obj):  # noqa N802
-        return DIRECTION(obj["_direction"])
+    def _decode_DIRECTION(self, obj: dict) -> DIRECTION:  # noqa: N802
+        return DIRECTION(obj['_direction'])
 
-    def decode_Graph(self, g_dict: dict[str, list | dict]) -> dict[str, list]:  # noqa N802
-        g = Graph(self.object_hook(g_dict["_direction"]))
-        g.nodes = self.object_hook(g_dict["_nodes"])
-        g.pipes = self.object_hook(g_dict["_pipes"])
+    def _decode_Graph(self, g_dict: dict[str, list | dict]) -> Graph:  # noqa: N802
+        g = Graph(self.object_hook(g_dict['_direction'])) # pyright: ignore[reportArgumentType]
+        g.nodes = self.object_hook(g_dict['_nodes']) # pyright: ignore[reportArgumentType]
+        g.pipes = self.object_hook(g_dict['_pipes']) # pyright: ignore[reportArgumentType]
+        g.qgis_fids = self.object_hook(g_dict['_qgis_fids']) # pyright: ignore[reportArgumentType]
         return g
+
+
+if __name__ == '__main__':
+    with open(r"C:/Users/holmest1/Greater Western Water/IP - Planning(Local) - Sewer/4. General/System Schematic/v2/data\\622d700622e88e2f.json") as jf:
+        j = json.load(jf, cls=ExtendedDecoder)