+++ gwwnetworktrace/plugin.py
@@ -112,16 +112,16 @@ class Plugin:
     def initGui(self) -> None:  # noqa N802
         """Create the menu entries and toolbar icons inside the QGIS GUI."""
         self.add_action(
-            icon_path="./resources/upstream.png",
-            text="Run upstream trace on selected feature",
-            callback=self.run_upstream,
+            icon_path="./resources/upstream_trace.png",
+            text="Run trace on selected feature/layer",
+            callback=self.run_trace,
             parent=iface.mainWindow(),
             add_to_toolbar=True,
         )
         self.add_action(
             icon_path="./resources/downstream.png",
-            text="Generate upstream graph",
-            callback=self.run_downstream,
+            text="Generate graph for selected layer",
+            callback=self.generate_graph,
             parent=iface.mainWindow(),
             add_to_toolbar=True,
         )
@@ -138,10 +138,10 @@ class Plugin:
         teardown_logger(Plugin.name)
         QgsApplication.processingRegistry().removeProvider(self.provider)
 
-    def run_upstream(self) -> None:
+    def run_trace(self) -> None:
         """Run method that performs all the real work"""
-        print("Run upstream trace")  # noqa: T201
+        print("Run network trace")  # noqa: T201
 
-    def run_downstream(self) -> None:
+    def generate_graph(self) -> None:
         """Run method that performs all the real work"""
-        print("Run downstream trace")  # noqa: T201
+        print("Generate network trace")  # noqa: T201