+++ CHANGELOG.md
@@ -1,5 +1,16 @@
-# CHANGELOG
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+- Select associated parcels
 
 
-
-###
+## [1.0.0] - 2024-07-12
+### Added
+- Initial release of GWW Network Trace plugin.
+- Ability to select upstream or downstream pipes from a single selected pipe.